# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
		"/Violation/allviolations": {
			"get": {
				"operationId": "GetAllViolations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"violations": {
											"items": {
												"properties": {
													"flagged": {
														"type": "string"
													},
													"location": {
														"type": "string"
													},
													"type": {
														"type": "string"
													},
													"plate": {
														"type": "string"
													},
													"date": {
														"type": "string",
														"format": "date-time"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"flagged",
													"location",
													"type",
													"plate",
													"date",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"violations"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/Violation/violationById": {
			"post": {
				"operationId": "GetViolationsById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"violation": {
													"properties": {
														"flagged": {
															"type": "string"
														},
														"location": {
															"type": "string"
														},
														"type": {
															"type": "string"
														},
														"plate": {
															"type": "string"
														},
														"date": {
															"type": "string",
															"format": "date-time"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"flagged",
														"location",
														"type",
														"plate",
														"date",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"violation"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/Violation/updateViolationById": {
			"put": {
				"operationId": "UpdateUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"violation": {}
											},
											"required": [
												"violation"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"location": {
										"type": "string"
									},
									"type": {
										"type": "string"
									},
									"plate": {
										"type": "string"
									},
									"date": {
										"type": "string",
										"format": "date-time"
									},
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/User/allusers": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"role": {
												"type": "string"
											},
											"password": {
												"type": "string"
											},
											"username": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"role",
											"password",
											"username",
											"name",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/User/userById": {
			"post": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"user": {
													"properties": {
														"role": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"username": {
															"type": "string"
														},
														"name": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"role",
														"password",
														"username",
														"name",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/User/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"user": {
													"properties": {
														"role": {
															"type": "string"
														},
														"username": {
															"type": "string"
														},
														"name": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"role",
														"username",
														"name",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/User/createUser": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"user": {
													"properties": {
														"role": {
															"type": "string"
														},
														"username": {
															"type": "string"
														},
														"name": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"role",
														"username",
														"name",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"role",
									"password",
									"username",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/User/updateUserById": {
			"put": {
				"operationId": "UpdateUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"user": {
													"properties": {
														"role": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"username": {
															"type": "string"
														},
														"name": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"role",
														"password",
														"username",
														"name",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"user"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/User/deleteUser": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"message": {},
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/hello": {
			"get": {
				"operationId": "SayHello",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/Report/createReport": {
			"post": {
				"operationId": "CreateReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"report": {
													"properties": {
														"includeViolationType": {
															"type": "boolean"
														},
														"includePlate": {
															"type": "boolean"
														},
														"violationId": {
															"type": "number",
															"format": "double"
														},
														"officerId": {
															"type": "number",
															"format": "double"
														},
														"dateCreated": {
															"type": "string",
															"format": "date-time"
														},
														"status": {
															"type": "string"
														},
														"content": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"includeViolationType",
														"includePlate",
														"violationId",
														"officerId",
														"dateCreated",
														"status",
														"content",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"report"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"includeViolationType": {
										"type": "boolean"
									},
									"includePlate": {
										"type": "boolean"
									},
									"officerId": {
										"type": "number",
										"format": "double"
									},
									"violationId": {
										"type": "number",
										"format": "double"
									},
									"status": {
										"type": "string"
									},
									"content": {
										"type": "string"
									}
								},
								"required": [
									"includeViolationType",
									"includePlate",
									"officerId",
									"violationId",
									"status",
									"content"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/Report/getReportById": {
			"post": {
				"operationId": "GetReportById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"reports": {
													"items": {
														"properties": {
															"includeViolationType": {
																"type": "boolean"
															},
															"includePlate": {
																"type": "boolean"
															},
															"violationId": {
																"type": "number",
																"format": "double"
															},
															"officerId": {
																"type": "number",
																"format": "double"
															},
															"dateCreated": {
																"type": "string",
																"format": "date-time"
															},
															"status": {
																"type": "string"
															},
															"content": {
																"type": "string"
															},
															"id": {
																"type": "number",
																"format": "double"
															}
														},
														"required": [
															"includeViolationType",
															"includePlate",
															"violationId",
															"officerId",
															"dateCreated",
															"status",
															"content",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"reports"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Local development server"
		},
		{
			"url": "https://traffic-violation-backend.onrender.com",
			"description": "Production server"
		}
	]
}